updated paths
