+++ src/auth/auth.module.ts
@@ -2,7 +2,8 @@ import { Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersRepository } from './user.repository';
+import { UsersRepository } from './repository/user.repository';
+import { RolesRepository } from './repository/role.repository';
 import { JwtModule } from '@nestjs/jwt';
 import { JwtStrategy } from './jwt.strategy';
 import { PassportModule } from '@nestjs/passport';
@@ -10,7 +11,7 @@ import { PassportModule } from '@nestjs/passport';
 @Module({
   controllers: [AuthController],
   imports: [
-    TypeOrmModule.forFeature([UsersRepository]),
+    TypeOrmModule.forFeature([UsersRepository, RolesRepository]),
     JwtModule.register({ secret: '1234', signOptions: { expiresIn: 3600 } }),
     PassportModule.register({ defaultStrategy: 'jwt' }),
   ],
