+++ src/auth/jwt.strategy.ts
@@ -3,8 +3,8 @@ import { PassportStrategy } from '@nestjs/passport';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { JwtPayload } from './jwt-payload.interface';
-import { User } from './user.entity';
-import { UsersRepository } from './user.repository';
+import { User } from './entity/user.entity';
+import { UsersRepository } from './repository/user.repository';
 
 export class JwtStrategy extends PassportStrategy(Strategy) {
   constructor(
